==> 01_Extraction.Rmd <==
---
title: "01 Data Extraction"
output: html_document
---

## Setup

```{r setup}
knitr::opts_chunk$set(warning = F, message = F)

packages <- c("tidyverse",
              "countrycode",
              "readxl")


pacman::p_load(packages, character.only = T)
```

## Pre-Existing Data

### GWIP

Breznau, Nate; Lanver, Felix, 2020, "Breznau and Lanver 2020 Global Work-Injury Policy Database (GWIP) Codebook.pdf", Global Work-Injury Policy Database (GWIP), <https://doi.org/10.7910/DVN/IVKYIE/EYP0PB>, Harvard Dataverse, V1.1

**Note** we use GWIP_v1.1 which is slightly updated from GWIP_v1.

Notes for the current replacement rate scores.
- These are for single workers if this can be calculated. Often the information is lacking to determine exactly what a worker gets depending on family situation
- For high demographic transition countries we assume two children, for low one. 
- If there are benefits that depend on multiple criteria that would adjust the amount upward, we usually add 5% to the base amount
- If there is a range depending on severity we usually take the average. For example, in The Netherlands the range is 70-100% of earnings, so we enter 85 (the midpoint). If 100% disability is 100% but then it goes down for less, and there are no other additional benefits we calculate 90.
- If there is a change within the first 6 months we make a calculation. For example, in Germany it is 100% for the first 6 weeks then 90 percent for the next 18, which equals 92.5 for the first 6 months. 
- Decentralized systems (by state or province), we try to take the average by state.

```{r gwip}
gwip <- read_csv(here::here("data", "gwip_v11.csv")) %>%
  mutate(ISO3c = countrycode(country_name, "country.name", "iso3c"),
         ISO3c = ifelse(country_name == "Republic of Vietnam", "VNM",
                        ifelse(country_name == "Serbia", "SRB",
                               ifelse(country_name == "Puerto Rico", "PRI", ISO3c))),
         labor_workinjury_firstins = ifelse(labor_workinjury_firstins == "none", NA, as.numeric(labor_workinjury_firstins)))

# there are some -99 that are missing values
gwip[gwip == -99] <- NA


```

### ILO & SSPW

The ILO released an impressive and comprehensive database of work-injury insurance coverage in 2014. It appears that this program was discontinued or at least has not been updated since then.

ILO. 2014. ["Global Programme Employment Injury Insurance and Protection | GEIP Data."](https://www.ilo.org/wcmsp5/groups/public/---ed_emp/---emp_ent/documents/publication/wcms_573083.pdf). Data scraped by Breznau, Nate and available in the `/data` folder of this project as `EIIP_2014.csv`.

However, after careful review of the data, 3 countries' scores are most certainly off by one decimal place. These are Niger, Sudan and Togo. 

This we can verify by looking at an even more impressive and most recently released World Social Protection Report (WSPR) from the ILO.

ILO. 2021. ["World Social Protection Report"](https://www.ilo.org/publications/flagship-reports/world-social-protection-report-2020-22-social-protection-crossroads-pursuit).

Users can otherwise find the original data in the files as `EIIP.csv` and `Social_Prot_Report_Tbl4.2_2020` which we laredy combined in the file `data/ilo_sspw_2020.csv`

We also hand coded the Social Security Programs Throughout the World (SSPW) 2018/2019 edition from https://www.ssa.gov/policy/docs/progdesc/ssptw/. We coded permanent and temporary replacement rates and their duration. More details available in the forthcoming technical paper. These data are also in the file `data/ilo_sspw_2020.csv`

```{r ilo_sspw}
ilo_sspw <- read.csv(here::here("data", "ilo_sspw_2020.csv")) %>%
  select(-country_name)
```



### SPIN - SIED

Only 37 countries. Use for Glolbal North point of comparison.

Social Insurance Entitlements Dataset (SIED)

Nelson, K., Fredriksson, D., Korpi, T., Korpi, W., Palme, J. and O. Sjöberg. 2020. The Social Policy Indicators (SPIN) database. International Journal of Social Welfare, 29(3), 285-289. <https://doi.org/10.1111/ijsw.12418>

|var_name|description|
|------|--------------------------------------|
|acovratl|Coverage ratio as proportion of labour force (ANOINSUR /ALABFORC)
|scovratl|Coverage ratio as proportion of labour force (SNOINSUR /SLABFORC)|
|scovratp|Coverage ratio as proportion of population (SNOINSUR / SPOP1564)|
|azrr26si|Standard net single worker replacement rate for total 26-week period of work accident benefit, excluding prior half-year's wage income from numerator and denominator| 
|azrr26fa|Standard net familied worker replacement rate for total 26-week period of work accident benefit, excluding prior half-year's wage income from numerator and denominator| 
|artsw26s|Standard gross single worker replacement rate for total 26-week period of work accident benefit| 
|artsw26f|Standard gross familied worker replacement rate for total 26-week period of work accident benefit| 

```{r sied}
sied <- read_csv(here::here("data", "SIED 1930–2020 (2023-06).csv")) %>%
  subset(year > 2000) %>%
  mutate(ISO3c = country) %>%
  select(ISO3c, year, acovratl, scovratl, scovratp, azrr26si, azrr26fa, az4ind, artsw26s, artsw26f) %>% 
  group_by(ISO3c) %>%
  mutate(Work_Inj_Cov_LF = max(acovratl, na.rm = T),
         Sick_Cov_LF = max(scovratl, na.rm = T),
         Sick_Cov_Pop = max(scovratp, na.rm = T),
         RR_Sing = max(azrr26si, na.rm = T),
         RR_Fam = max(azrr26fa, na.rm = T),
         RR_Avg_Sing = max(az4ind, na.rm = T),
         RRgross_Sing = max(artsw26s, na.rm = T),
         RRgross_Fam = max(artsw26f, na.rm = T)) %>%
  select(-c(acovratl, scovratl, scovratp, azrr26si, azrr26fa, az4ind, artsw26s, artsw26f)) %>%
  summarise_all(mean, na.rm = T) %>%
  mutate(year = 2020) %>%
  ungroup()
```

```{r corr}
siedcor <- read_csv(here::here("data", "SIED 1930–2020 (2023-06).csv")) %>%
  subset(year > 2000) %>%
  select(azrr26si, azrr26fa, artsw26s, artsw26f) %>%
  cor(., use = "pairwise.complete.obs") %>%
  round(., 2)

siedcor
```


### CWED2

Nelson, Kenneth, Daniel Fredriksson, Tomas Korpi, Walter Korpi, Joakim Palme, and Ola Sjöberg. 2020. “The Social Policy Indicators (SPIN) Database.” International Journal of Social Welfare 29(3):285–89. doi: 10.1111/ijsw.12418.


Does not have work-injury policy, only sickness insurance. This blurs the health care and work-injury systems in case they are separate. THerefore, we leave this here as evidence that we considered it - it is afterall a central dataset in welfare state research.

```{r cwed}
#cwed2 <- read_csv(here::here("data", "cwed-subset.csv"))
```



### Disability RR ILO

This is a rough measure to help with imputation. It calculates the ratio of the average income of persons with a disability to those without.

https://www.ilo.org/shinyapps/bulkexplorer44/?lang=en&id=EAR_4MTH_SEX_DSB_CUR_NB_A

*Average monthly earnings of employees by sex and disability status -- Annual*

With the aim of promoting international comparability, statistics presented on ILOSTAT are based on standard international definitions wherever feasible and may differ from official national figures. This series is based on the 13th ICLS definitions. For time series comparability, it includes countries that have implemented the 19th ICLS standards, for which data are also available in the Work Statistics -- 19th ICLS (WORK) database. The earnings of employees relate to the gross remuneration in cash and in kind paid to employees, as a rule at regular intervals, for time worked or work done together with remuneration for time not worked, such as annual vacation, other type of paid leave or holidays. Earnings exclude employers' contributions in respect of their employees paid to social security and pension schemes and also the benefits received by employees under these schemes. Earnings also exclude severance and termination pay. This is a harmonized series: (1) data reported as weekly, monthly and yearly are converted to hourly using data on average weekly hours if available; and (2) data are converted to U.S. dollars as the common currency, using exchange rates or using purchasing power parity (PPP) rates for private consumption expenditures. The latter series allows for international comparisons by taking account of the differences in relative prices between countries. The term disability, as defined in the International Classification of Functioning, Disability and Health (ICF), is used as an umbrella term, covering impairments, activity limitations, and participation restrictions. For measurement purposes, a person with disability is defined as a person who is limited in the kind or amount of activities that he or she can do because of ongoing difficulties due to a long-term physical condition, mental condition or health problem. For more information, refer to the Disability Labour Market Indicators (DLMI) database description.

Subject: Earnings
Database: Disability Labour Market Indicators (DLMI)
Id: EAR_4MTH_SEX_DSB_CUR_NB_A
Time period: 1996 - 2023
Last updated: 04/02/2024 06:45:18


```{r disrr}
dis_ilo <- read_csv(here::here("data", "EAR_4MTH_SEX_DSB_CUR_NB_A-filtered-2024-02-05.csv")) %>%
  subset(classif2.label == "Currency: 2017 PPP $" & sex.label == "Sex: Total") %>%
  mutate(ISO3c = countrycode::countrycode(ref_area.label, "country.name", "iso3c"),
         year = time) %>%
  select(ISO3c, year, classif1.label, obs_value) %>%
  subset(classif1.label != "Disability status (Aggregate): Total") %>%
  group_by(ISO3c) %>%
  mutate(maxyear = max(year, na.rm = T),
         year_obs = year,
         year = 2020) %>%
  ungroup() %>%
  subset(year_obs == maxyear) %>%
  select(-maxyear) %>%
  group_by(ISO3c) %>%
  mutate(Dis_RR = round((min(obs_value)/max(obs_value))*100,3),
         minn = min(obs_value)) %>%
  subset(obs_value != minn, select = c(ISO3c, year, Dis_RR, year_obs)) %>%
  ungroup()
```

### Social Indicators

#### GDP Maddison

Includes population

https://www.rug.nl/ggdc/historicaldevelopment/maddison/releases/maddison-project-database-2020?lang=en

```{r gdp}
gdp <- read_excel(here::here("data", "mpd2020.xlsx"), sheet = "Full data") %>%
  mutate(ISO3c = countrycode,
         year_obs = year,
         year = 2020) %>%
  subset(year_obs == 2018, select = c(ISO3c, year, gdppc, pop, year_obs))
```


#### Gini Solt

Solt, Frederick. 2020. ‘The Standardized World Income Inequality Database, Versions 8-9’. https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/LM4OWF

Think about calculating a redistribution score from gini_disp and gini_mkt.

```{r gini}
load(here::here("data", "swiid8_3.rda"))

gini <- swiid_summary %>% 
   mutate(ISO3c = countrycode(country, "country.name","iso3c")) %>% 
   select(ISO3c, gini_disp, gini_mkt, year) %>% 
   group_by(ISO3c) %>% 
   mutate(year_max = max(year),
          year_filter = year-year_max,
          year_obs = year,
          year = 2020) %>%
   subset(year_filter > -5, select = -c(year_max, year_filter)) %>%
   summarise_all(mean, na.rm = T) %>% 
   subset(!is.na(ISO3c))

rm(swiid, swiid_summary)
```

### LFP ILO

https://www.ilo.org/shinyapps/bulkexplorer14/?lang=en&segment=indicator&id=EAP_2WAP_SEX_AGE_RT_A

Labour force participation rate by sex, age and rural / urban areas -- ILO modelled estimates, Nov. 2021 (%) -- Annual

Imputed observations are not based on national data, are subject to high uncertainty and should not be used for country comparisons or rankings. This series is based on the 13th ICLS definitions. The labour force participation rate is the labour force as a percent of the working-age population. The labour force is the sum of all persons of working age who are employed and those who are unemployed. For more information, refer to the ILO Modelled Estimates (ILOEST) database description.

Subject: Labour force
Database: ILO Modelled Estimates (ILOEST) | Rural and Urban Labour Markets (RURBAN)
Id: EAP_2WAP_SEX_AGE_GEO_RT_A
Time period: 2005 - 2020
Last updated: 27/06/2022 12:21:39
Number of records: 118800


```{r lfp}


lfp_total <- read_csv(here::here("data", "EAP_2WAP_SEX_AGE_GEO_RT_A-filtered-2024-02-06.csv")) %>%
  subset(sex.label == "Sex: Total" & time > 2015 & classif1.label == "Age (Youth, adults): 25+" & classif2.label == "Area type: National") %>% 
  mutate(ISO3c = countrycode::countrycode(ref_area.label, "country.name", "iso3c"),
         ISO3c = ifelse(ref_area.label == "Türkiye", "TUR", ISO3c),
         lfp_total = obs_value) %>%
  select(ISO3c, lfp_total) %>%
  group_by(ISO3c) %>%
  summarise_all(mean, na.rm = T) %>%
  mutate(year = 2020) %>%
  ungroup()

lfp_female <- read_csv(here::here("data", "EAP_2WAP_SEX_AGE_GEO_RT_A-filtered-2024-02-06.csv")) %>%
  subset(sex.label == "Sex: Female" & time > 2015 & classif1.label == "Age (Youth, adults): 25+" & classif2.label == "Area type: National") %>% 
  mutate(ISO3c = countrycode::countrycode(ref_area.label, "country.name", "iso3c"),
         ISO3c = ifelse(ref_area.label == "Türkiye", "TUR", ISO3c),
         lfp_female = obs_value) %>%
  select(ISO3c, lfp_female) %>%
  group_by(ISO3c) %>%
  summarise_all(mean, na.rm = T) %>%
  mutate(year = 2020) %>%
  ungroup()
```

## Merge All Data

### Dataframe

```{r merge}
# Setup dataframe
df <- as.data.frame(gwip)

# Add collected data

df <- df %>%
  left_join(ilo_sspw, by = "cow_code") %>%
  left_join(sied, by = "ISO3c") %>%
  left_join(subset(dis_ilo, select = c(Dis_RR, ISO3c)), by = "ISO3c") %>%
  left_join(subset(gini, select = c(gini_disp, gini_mkt, ISO3c))) %>%
  left_join(subset(lfp_female, select = c(lfp_female, ISO3c)), by = "ISO3c") %>%
  left_join(subset(lfp_total, select = c(lfp_total, ISO3c)), by = "ISO3c") %>%
  left_join(subset(gdp, select = c(gdppc, pop, ISO3c)), by = "ISO3c") %>%
  subset(!is.na(ISO3c))
```


### Fixes

#### Background

We add colonial legacy (including Russia/USSR), an indicator of being in the so called 'Global South' and UN region. 

```{r background}

background <- read.csv(here::here("data", "background.csv"))

df <- df %>%
  left_join(background, by = "cow_code")
```


#### GDP

This is simply "GDP per capita 2020" from Google which draws its default data from the World Bank. It is not perfectly compatible with Maddisson, but very close conceptually. I round to flat 1,000s so they are easy to notice as imputed. I also round Monaco and Leichtenstein down to 100000 for parsimony. 

```{r gdp_fix}
miss_gdp <- df$ISO3c[is.na(df$gdppc)]
miss_gdp_name <- countrycode(df$ISO3c[is.na(df$gdppc)], "iso3c", "country.name")

df <- df %>%
  mutate(gdppc = case_when(
    ISO3c == "BHS" ~ 23000,
    ISO3c == "GRD" ~ 8000,
    ISO3c == "VCT" ~ 8000,
    ISO3c == "ATG" ~ 15000,
    ISO3c == "KNA" ~ 19000,
    ISO3c == "BLZ" ~ 5000,
    ISO3c == "GUY" ~ 7000,
    ISO3c == "SUR" ~ 5000,
    ISO3c == "MCO" ~ 100000,
    ISO3c == "LIE" ~ 100000,
    ISO3c == "AND" ~ 37000,
    ISO3c == "SMR" ~ 45000,
    ISO3c == "SOM" ~ 500,
    ISO3c == "SSD" ~ 500,
    ISO3c == "BTN" ~ 3000,
    ISO3c == "BRN" ~ 27000,
    ISO3c == "PNG" ~ 2000,
    ISO3c == "VUT" ~ 3000,
    ISO3c == "SLB" ~ 2000,
    ISO3c == "KIR" ~ 1000,
    ISO3c == "FJI" ~ 5000,
    ISO3c == "MHL" ~ 5000,
    ISO3c == "PLW" ~ 14000,
    ISO3c == "FSM" ~ 4000,
    ISO3c == "WSM" ~ 4000,
    .default = gdppc
  ),
  gdppc_k = gdppc/1000)
```

#### Var Labels

```{r labels}
df_labels <- as.data.frame(colnames(df)) %>%
  mutate(variables = `colnames(df)`,
         labels = case_when(
           variables == "cow_code" ~ "Correlates of War 3-digit alpha code",
           variables == "ISO3c" ~ "International Standards of Classification Country Alpha-3 Categorization",
           variables == "year" ~ "Year for merging (observations are 2020 or closest)",
           variables == "Work_Inj_Cov_LF" ~ "(acovratl) Coverage ratio as proportion of labour force (ANOINSUR /ALABFORC), source, SIED 2020",
           variables == "Sick_Cov_LF" ~ "(scovratl) Coverage of the labor force for work-injury policy, source = SIED 2020",
           variables == "Sick_Cov_Pop" ~ "(scovratp) Coverage of the entire adult population for work-injury policy, source = SIED 2020",
           variables == "RR_Sing" ~ "(azrr26si) Standard net single worker replacement rate for total 26-week period of work accident benefit, excluding prior half-year's wage income from numerator and denominator, source SIED 2020",
           variables == "RR_Fam" ~ "(azrr26fa) Standard net familied worker replacement rate for total 26-week period of work accident benefit, excluding prior half-year's wage income from numerator and denominator, source = SIED 2020",
           variables == "RR_Avg_Sing" ~ "(az4ind) (= a1stners+a1stnerf+azrr26si+azrr26fa)/4. Average of four components: a single person and a four-person family, for first week after waiting days and 26 weeks with benefits, source = SIED 2020",
           variables == "replacement_rate_perm" ~ "The replacement rate for a permanent disability for the first 12 months of work incapacity, compared to previous wages, or average wage if so calculated. Coded from Social Policy Programs Throughout the World, or best available info. Source = GWIP v2.0",
           variables == "duration_perm" ~ "The length of time a person who has a permanent disability caused by a work-injury is covered, source = GWIP v2.0",
           variables == "replacement_rate_temp" ~ "The replacement rate for a temporary incapacity to work compared to previous wages, or average wage if so calculated, in the first 6 months of incapacity, source = GWIP v2.0",
           variables == "duration_temp" ~ "The length of time a perosn who has a temporary incapacity to work can recieve benefits. Unlimited = until healed. Source = GWIP v2.0",
           variables == "qualifying_period" ~ "If any duration of work or other qualifications other than a work-injury have an impact on the rate or duration of benefits, source = GWIP v2.0",
           variables == "Dis_RR" ~ "The ratio of income for a person with disability to one without, source = ILO 2020",
           variables == "gini_disp" ~ "Disposable Gini. The income inequality Gini after taxes and transfers, source = Solt 2020",
           variables == "gini_mkt" ~ "Market Gini. The income inequality Gini before taxes and transfers, source = ILO 2020",
           variables == "workinjury_coverage_GEIP" ~ "Percent of labor force de Facto covered by work-injury policy, source Global Empoyment Injury Protection ILO 2014",
           variables == "workinjury_coverage_WSPR" ~ "Percent of labor force de Factor covered by work-injury policy, source ILO World Social Protection Report 2020-2022",
           variables == "`Program Type`" ~ "Program type based on GEIP 2014, source ILO 2014",
           variables == "lfp_female" ~ "Labor force participation among adult females, source = ILO 2020",
           variables == "lfp_total" ~ "Labor force participation among all adults, source = ILO 2020",
           variables == "gdppc" ~ "Gross Domestic Product per capita, source = Maddison 2020 & 26 very small countries Google (World Bank)",
           variables == "pop" ~ "Population, source = Maddison 2020",
           variables == "labor_workinjury_firstlaw" ~ "First work-injury law, source = GWIP v1.1",
           variables == "labor_workinjury_firstins" ~ "First law including risk-pooling (insurance or fund), source = GWIP v1.1",
           variables == "gdppc_k" ~ "GDP per capita in 1,000 US$ 2020, source = Maddisson 2020 & 26 countries Google (World Bank)",
           variables == "colonial_history" ~ "Colonial Legacy, source = https://gist.github.com/caluchko with author updates for smaller countries",
           .default = NA)) %>%
  subset(., select = c(variables, labels))

write_csv(df, here::here("data", "df.csv"))
write_csv(df_labels, here::here("data", "df_labels.csv"))
```


## Environment

### Colophon

```{r colophon}
sessionInfo()
```

### Citations

```{r citations}
packages %>%
  map(citation) %>%
  print(style = "text")
```
