==> 02_Measurement.Rmd <==
---
title: "02_Data_Measurments"
author: "Nate Breznau"
date: "`r Sys.Date()`"
output: html_document
---

## Setup

Datafiles df.csv and df_labels.csv created in the file 01_Data_Extraction.Rmd.

```{r setup}
knitr::opts_chunk$set(warning = F, message = F)

packages <- c("tidyverse",
              "countrycode",
              "ggplot2",
              "ggrepel",
              "ggpubr",
              "sjPlot",
              "ragg")

pacman::p_load(packages, character.only = T)

df <- as.data.frame(read_csv(here::here("data", "df.csv")))
df_labels <- as.data.frame(read_csv(here::here("data", "df_labels.csv")))

#fix region
df <- df %>%
  mutate(Region_factor = as.factor(Region))
```

## Coverage

### Coverage Adjust

We have 2 ILO variables for coverage, as they differ sometimes between the GEIP and WSPR, we take the average. WSPR must be divided by 100 as it is in percentage and GEIP in ratio.

The most comprehensive data relates to coverage, therefore I start here to develop improvements in measurement and impute missing values. Then I simply call it the ILO data from here on.

```{r iloavg}
df <- df %>%
  mutate(workinjury_coverage_ILO = rowMeans(cbind(workinjury_coverage_GEIP, workinjury_coverage_WSPR/100), na.rm = T),
         workinjury_coverage_ILO = ifelse(workinjury_coverage_ILO == "NaN", NA, workinjury_coverage_ILO)) %>%
  ungroup()

#length(df$workinjury_coverage_ILO[!is.na(df$workinjury_coverage_ILO)])
```


A first issue relates to some incompatibilities of the SIED and the GEIP. Although the former is from 2020 and the latter from 2014, the legal coverage of the labor force should not be so different. It has not changed in most rich countries. The correlation is slightly above 0.5. 

### Fig 7. Comparing SIED & GEIP

```{r cov}
#calculate correlation and p
cov_cor <- subset(df, select = c(workinjury_coverage_ILO, Work_Inj_Cov_LF))

cov_cor_R <- cov_cor %>%
  cor(., use = "pairwise.complete.obs") %>%
  round(.,2)[2]

cov_cor_P <- paste0("R = ", round(cov_cor_R,2)[1], ", (", round(cor.test(as.numeric(unlist(cov_cor[,1])),as.numeric(unlist(cov_cor[,2])))$p.value,3)[1], ")")

agg_png(here::here("results", "Fig7.png"), res = 144, width = 800, height = 550)

df %>%
  #subset(ISO3c != "ROU") %>%
ggplot(aes(workinjury_coverage_ILO, Work_Inj_Cov_LF)) +
  geom_smooth(method = "lm", color = "#472D7BFF", size = 0.4) +
  annotate(geom = "text", 
           size = 3, 
           color = "#472D7BFF",
           x = 0.95,
           y = 0.6,
           label = cov_cor_P) +
  xlab("ILO Data") +
  ylab("SIED Data") +
  geom_text_repel(aes(label = ISO3c), size = 3) +
  xlim(.6,1) +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results", "Fig7.png"))
  
  
```


### Fig Extra. Disability Pay Gap

The idea is to consider increasing the predictive power of the GEIP through other variables. The first candidate is the difference in average income between normal workers and disabled persons. This should pick up on both the generosity and especially the coverage of disability insurance, which is linked to work-injury policy. But the fit is weak at best.

```{r lf}
df %>%
  ggplot(aes(workinjury_coverage_ILO, Dis_RR)) +
  geom_text_repel(aes(label = ISO3c), size = 2) +
    stat_cor(method = "pearson", 
           size = 2.5, 
           color = "#472D7BFF",
           label.x = .8,
           label.y = 0.6) +
  theme_classic()
```

### Tbl Extra. Correlation Matrix

In further seeking to understand the differences in scores in coverage measures we present a full correlation matrix with all the data at hand that includes countries both within and without of the SIED, in pairwise correlation format. 

```{r pairwise_corr}
cor1 <- df %>%
  mutate(diff_sied_geip = workinjury_coverage_ILO - Work_Inj_Cov_LF) %>%
  subset(select = c(workinjury_coverage_ILO, Work_Inj_Cov_LF, diff_sied_geip, Sick_Cov_LF, Sick_Cov_Pop, gdppc_k, pop, lfp_total, lfp_female, Dis_RR, replacement_rate_perm_sspw, replacement_rate_temp_sspw, labor_workinjury_firstlaw, labor_workinjury_firstins)) %>%
cor(., use = "pairwise.complete.obs") %>%
  round(.,3)

cor1[upper.tri(cor1)] <- ""

cor1


```

### Prediction Model


#### Base Regressions

I constrain the intercept to zero on the assumption that GDP is crucial for coverage rates. As the SIED data are missing the entire Global South, this assumption allows for more accurate imputation.

Note that the r-squared values are wrong here when removing the intercept. 

```{r predict}
m1 <- lm(Work_Inj_Cov_LF ~ workinjury_coverage_ILO + 0, data = df)
m2 <- lm(Work_Inj_Cov_LF ~ workinjury_coverage_ILO + 0 + gdppc_k , data = df)
m3 <- lm(Work_Inj_Cov_LF ~ workinjury_coverage_ILO + 0 + gdppc_k + lfp_total , data = df)
m4 <- lm(Work_Inj_Cov_LF ~ workinjury_coverage_ILO + 0 + gdppc_k + labor_workinjury_firstins, data = df)
m5 <- lm(Work_Inj_Cov_LF ~ workinjury_coverage_ILO + 0 + lfp_total + labor_workinjury_firstins, data = df)

tab_model(m1,m2,m3,m4,m5, p.style = "stars", show.ci = F, show.r2 = F)

reg1 <- matrix(nrow = 5, ncol = 11)
colnames(reg1) <- c("Predictor", "m1_b", "m1_p", "m2_b", "m2_p", "m3_b", "m3_p", "m4_b", "m4_p", "m5_b", "m5_p")
reg1[,1] <- c("GEIP_Coverage", "GDPpc", "LFP", "First Law", "R2")
reg1[1,2:3] <- round(summary(m1)$coefficients[,c(1,4)],3)
reg1[1:2,4:5] <- round(summary(m2)$coefficients[,c(1,4)],3)
reg1[1:3,6:7] <- round(summary(m3)$coefficients[,c(1,4)],3)
reg1[c(1,2,4),8:9] <- round(summary(m4)$coefficients[,c(1,4)],3)
reg1[c(1,3,4),10:11] <- round(summary(m5)$coefficients[,c(1,4)],3)
  
# Predicted values 
df$Work_Inj_Cov_Pred <- predict.lm(m2, newdata = df)

# get correct r-squared
mcors <- df
mcors$m1_pred <- predict.lm(m1, newdata = mcors)
mcors$m3_pred <- predict.lm(m3, newdata = mcors)
mcors$m4_pred <- predict.lm(m4, newdata = mcors)
mcors$m5_pred <- predict.lm(m5, newdata = mcors)

mcors <- mcors %>%
  mutate(!is.na(Work_Inj_Cov_LF)) %>%
  select(Work_Inj_Cov_LF, m1_pred, Work_Inj_Cov_Pred, m3_pred, m4_pred, m5_pred) %>%
  cor(., use = "pairwise.complete.obs")

mcors <- as.data.frame(mcors[2:6,1])

colnames(mcors) <- "R"

mcors <- mcors %>%
  mutate(R2 = R*R)

reg1[5,c(2,4,6,8,10)] <- round(mcors$R2,2)

reg1[is.na(reg1)] <- ""

write_csv(as.data.frame(reg1), here::here("results", "Tbl1.csv"))

print(as.table(reg1))


```

#### Residuals

The prediction is of high quality.
```{r resid}
# Coverage shall not be over 1.0
df <- df %>%
  mutate(Work_Inj_Cov_Pred = ifelse(Work_Inj_Cov_Pred > 1, 1, Work_Inj_Cov_Pred))

df %>%
  ggplot(aes(Work_Inj_Cov_LF, Work_Inj_Cov_Pred)) +
  geom_text_repel(aes(label = ISO3c), size = 2) +
  geom_abline(a = 0, slope = 1) +
  theme_classic()
```



### Fig 8. Plot Predictions


```{r plotted}
agg_png(here::here("results", "Fig8.png"), res = 144, width = 800, height = 550)

df %>%
  mutate(Work_Inj_Cov_Pred_Only = ifelse(is.na(Work_Inj_Cov_LF), Work_Inj_Cov_Pred, NA)) %>%
  ggplot() +
  geom_abline(a = 0, slope = 1, size = 1, color = "#472D7BFF") +
  geom_point(aes(workinjury_coverage_ILO, Work_Inj_Cov_Pred_Only), color = "#B8DE29FF", size = 1.5) +
  geom_point(aes(workinjury_coverage_ILO, Work_Inj_Cov_LF), color = "#33638DFF", size = 1.5) +
  annotate("text", 
           x = 0.6,
           y = 1,
           label = "Original\nSIED values",
           color = "#33638DFF",
           hjust = 1,
           size = 3.1) +
  annotate("text", 
           x = 0.38,
           y = 0.55,
           label = "Predicted\nfrom ILO",
           color = "#B8DE29FF",
           hjust = 1,
           size = 3.1) +
  xlab("ILO Data") +
  ylab("Harmonized Values\nStochastic Prediction") +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results", "Fig8.png"))
```

## Replacement Rates (Generosity)

### Fig Extra. Replacement Rates GEIP & SIED. RR Temp, Single Worker

Temp is the replacement rate for the first 6 months of disability resulting from a work-injury for a worker, with one child where applicable. 

The hand coded data here, mostly from the ISSA, track the single worker replacement rate.

```{r gen}
cov_cor <- subset(df, select = c(replacement_rate_temp_sspw, RR_Sing))

cov_cor_R <- cov_cor %>%
  cor(., use = "pairwise.complete.obs") %>%
  round(.,2)[2]

cov_cor_P <- paste0("R = ", round(cov_cor_R,2)[1], ", (", round(cor.test(unlist(cov_cor[,1]),unlist(cov_cor[,2]))$p.value,3)[1], ")")

agg_png(here::here("results", "Fig9.png"), res = 144, width = 800, height = 550)

df %>%
ggplot(aes(replacement_rate_temp_sspw/100, RR_Sing)) +
  geom_smooth(method = "lm", color = "#472D7BFF", size = 0.4, fullrange = T) +
  annotate(geom = "text", 
           size = 3, 
           color = "#472D7BFF",
           x = 0.85,
           y = 0.3,
           label = cov_cor_P) +
  geom_text_repel(aes(label = ISO3c), size = 2.5) +
  xlab("Replacement Rate Temp (GWIP)") +
  ylab("Replacement Rate Single (SIED)") +
  xlim(.2,1.1) +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results", "Fig9.png"))
```

### Fig Extra. Replacement Rates GEIP & SIED. RR Temp, Average Single Worker

```{r gen2}
df %>%
ggplot(aes(replacement_rate_temp_sspw/100, RR_Avg_Sing)) +
  geom_smooth(method = "lm", color = "#472D7BFF", size = 0.4) +
  stat_cor(method = "pearson", 
           size = 2.5, 
           color = "#472D7BFF",
           label.x = .9,
           label.y = 0.4
           ) +
  geom_text_repel(aes(label = ISO3c), size = 2.5) +
  xlim(.2,1) +
  theme_classic()
```

### Fig Extra. Replacement Rates GEIP & SIED. RR Temp, Family

Tracks family as well. 

```{r gen3}
df %>%
ggplot(aes(replacement_rate_temp_sspw/100, RR_Fam)) +
  geom_smooth(method = "lm", color = "#472D7BFF", size = 0.4) +
  stat_cor(method = "pearson", 
           size = 2.5, 
           color = "#472D7BFF",
           label.x = .9,
           label.y = 0.45
           ) +
  geom_text_repel(aes(label = ISO3c), size = 2.5) +
  xlim(.2,1) +
  theme_classic()
```


### Tbl Extra. Correlation Matrix RRs

```{r}
cor2 <- df %>%
  mutate(diff_sied_gwip = replacement_rate_temp_sspw - RR_Avg_Sing) %>%
  subset(select = c(replacement_rate_temp_sspw, RR_Avg_Sing, diff_sied_gwip, Sick_Cov_LF, Sick_Cov_Pop, gdppc_k, pop, lfp_total, lfp_female, Dis_RR, replacement_rate_perm_sspw, labor_workinjury_firstlaw, labor_workinjury_firstins)) %>%
cor(., use = "pairwise.complete.obs") %>%
  round(.,3)

cor2[upper.tri(cor2)] <- ""

cor2
```
### Prediction Model

#### Base Regressions

Here it does not make much sense to constrain the intercept to 0, because replacement rates are not expected to be as much of a product of industrialization as coverage. This means that replacement_rate_temp alone best predicts RR_Sing.

There are a few exceptions in SDN and SSD the replacement rate is determined to be zero in our data but ends up near 0.53 because of the intercept. This is a massive outlier, and therefore we simply recode it to be zero.

```{r regs_rr}

m11 <- lm(RR_Sing ~ replacement_rate_temp_sspw, data = df)
m12 <- lm(RR_Sing ~ replacement_rate_temp_sspw + gdppc_k, data = df)
m13 <- lm(RR_Sing ~ replacement_rate_temp_sspw + lfp_total, data = df)
m14 <- lm(RR_Sing ~ replacement_rate_temp_sspw + pop, data = df)
m15 <- lm(RR_Sing ~ replacement_rate_temp_sspw + labor_workinjury_firstins, data = df)

tab_model(m11,m12,m13,m14,m15, digits = 3, p.style = "stars", show.ci = F, file = here::here("results", "Tbl4.doc"))

  
# Valid predictions for Global North
df$RR_Sing_Pred <- predict.lm(m12, newdata = df)

# fix SDN
df <- df %>%
  mutate(RR_Sing_Pred = ifelse(ISO3c == "SDN", 0.1, RR_Sing_Pred),
         RR_Sing_Pred = ifelse(ISO3c == "SSD", 0.1, RR_Sing_Pred))

```

#### Residuals

The prediction is of high quality.

```{r resid_rr}

df %>%
  ggplot(aes(RR_Sing, RR_Sing_Pred)) +
  geom_text_repel(aes(label = ISO3c), size = 2) +
  geom_abline(a = 0, slope = 1) +
  theme_classic()
```


#### Fig 10. Plot Predictions


```{r plotted_RR}
agg_png(here::here("results", "Fig10.png"), res = 144, width = 800, height = 550)

df %>%
  mutate(RR_Sing_Pred_Only = ifelse(is.na(RR_Sing), RR_Sing_Pred, NA),
         RR_Sing_I = ifelse(is.na(RR_Sing), RR_Sing_Pred, RR_Sing)) %>%
  ggplot() +
  geom_smooth(aes(replacement_rate_temp_sspw/100, RR_Sing_I), method = "lm", se = F, size = 1, color = "#472D7BFF") +
  geom_point(aes(replacement_rate_temp_sspw/100, RR_Sing_Pred_Only), color = "#B8DE29FF", size = 1.5) +
  geom_point(aes(replacement_rate_temp_sspw/100, RR_Sing), color = "#33638DFF", size = 1.5) +
    annotate("text", 
           x = 0.75,
           y = 0.57,
           label = "Original\nSIED values",
           color = "#33638DFF",
           hjust = 0,
           size = 3.1) +
  annotate("text", 
           x = 0.3,
           y = 0.97,
           label = "Predicted\nfrom GWIP",
           color = "#B8DE29FF",
           hjust = 0,
           size = 3.1) +
  xlab("Replacement Rate Temp (GWIP)") +
  ylab("Replacement Rate SIED Harmonized") +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results", "Fig10.png"))

```

## Calculate Values

### SIED imputed baseline

```{r sied}
df <- df %>%
  mutate(Work_Inj_Cov_LF_imputed = ifelse(is.na(Work_Inj_Cov_LF), Work_Inj_Cov_Pred, Work_Inj_Cov_LF),
         RR_Sing_imputed = ifelse(is.na(RR_Sing), RR_Sing_Pred, RR_Sing))
```

### SIED imputed full

Now we make a larger leap into imputing countries for which there are no data in either the SIED, GWIP or GEIP

#### Coverage

```{r fullimp_cov}
# mean impute lfp_total for prediction purposes

df <- df %>%
  mutate(lfp_total_I = ifelse(is.na(lfp_total), mean(df$lfp_total, na.rm = T), lfp_total))

m21 <- lm(Work_Inj_Cov_Pred ~ gdppc_k, data = df)
m22 <- lm(Work_Inj_Cov_Pred ~ gdppc_k + lfp_total_I, data = df)
m23 <- lm(Work_Inj_Cov_Pred ~ gdppc_k + labor_workinjury_firstlaw, data = df)
m24 <- lm(Work_Inj_Cov_Pred ~ gdppc_k + labor_workinjury_firstlaw + lfp_total_I, data = df)

tab_model(m21,m22,m23,m24, show.ci = F, digits = 3, p.style = "stars", file = here::here("results", "Tbl3.doc"))

df$x <- predict.lm(m22, newdata = df)

# shall not be over 1.0 for coverage
df <- df %>%
  mutate(Work_Inj_Cov_LF_imputed_full = ifelse(is.na(Work_Inj_Cov_LF_imputed), x, Work_Inj_Cov_LF_imputed),
         Work_Inj_Cov_LF_imputed_full = ifelse(Work_Inj_Cov_LF_imputed_full > 1, 1, Work_Inj_Cov_LF_imputed_full)) %>%
  select(-c(x,Work_Inj_Cov_Pred,RR_Sing_Pred))

df$ISO3c[is.na(df$Work_Inj_Cov_LF_imputed_full)]

```

#### Generosity

##### Regressions
Tested gdppc & LFP, no effects

```{r fullim_gen}
# Polynesia does not work as a single region, recode to micronesia for the regression

df <- df %>%
  mutate(Region_factor = case_match(Region_factor,
                                      "Polynesia" ~ "Micronesia",
                                      .default = Region_factor))

m31 <- lm(RR_Sing_imputed ~ gini_disp, data = df)
m32 <- lm(RR_Sing_imputed ~ Region_factor, data = df)
m33 <- lm(RR_Sing_imputed ~ gini_disp + Region_factor, data = df)
m34 <- lm(RR_Sing_imputed ~ labor_workinjury_firstlaw*Region_factor, data = df)
m35 <- lm(RR_Sing_imputed ~ as.factor(colonial_history) + Region_factor, data = df)


tab_model(m31,m32,m33,m34,m35, digits = 3, show.ci = F, p.style = "stars", file = here::here("results", "Tbl5.doc"))



# The regression results are unclear as to which is the better predictor
df$xx <- predict.lm(m32, newdata = df)
df$xxx <- predict.lm(m34, newdata = df)

# then recode back for an accurate dataset in the end

df <- df %>%
  mutate(Region_factor = case_match(ISO3c,
                                    "PYF" ~ "Polynesia",
                                    .default = Region_factor))

df <- df %>%
  mutate(RR_Sing_imputed_full_32 = ifelse(is.na(RR_Sing_imputed), xx, RR_Sing_imputed),
         RR_Sing_imputed_full_34 = ifelse(is.na(RR_Sing_imputed), xxx, RR_Sing_imputed)) %>%
  select(-c(xx,xxx))
```

##### Comparison of Results

m32 and m34 have potentially useful prediction results, compare them to see how different they really are. m34 leads to some peculiar outliers. Given that none of the coefficients are significant and the intercept behaves oddly it is best to use m32

```{r compare}

agg_png(filename = here::here("results", "Fig11.png"), res = 144, width = 750, height = 500)

df %>%
  ggplot(aes(y = RR_Sing_imputed_full_32, x = RR_Sing_imputed_full_34)) +
  geom_abline(intercept = 0, slope = 1) +
  geom_point() +
  xlab("Model 34 imputation") +
  ylab("Model 32 imputation") +
  theme_classic()

dev.off()

df <- df %>%
  mutate(RR_Sing_imputed_full =  RR_Sing_imputed_full_32) %>%
  select(-c(RR_Sing_imputed_full_32, RR_Sing_imputed_full_34))

knitr::include_graphics(here::here("results", "Fig11.png"))

```

## Save GWIP

### Generate

```{r gwip20}
gwip20 <- df %>%
  mutate(cow_code = countrycode(ISO3c, "iso3c", "cown"),
         cow_code = ifelse(ISO3c == "PRI", 6, cow_code),
         cow_code = ifelse(ISO3c == "VNM", 817, cow_code),
         cow_code = ifelse(ISO3c == "SRB", 345, cow_code),
         workinjury_coverage_SIED_harmo = round(Work_Inj_Cov_LF_imputed,3),
         workinjury_coverage_SIED_harmo_full = round(Work_Inj_Cov_LF_imputed_full,3),
         workinjury_replacement_rate_single_SIED_harmo = round(RR_Sing_imputed,3),
         workinjury_replacement_rate_single_SIED_harmo_full = round(RR_Sing_imputed_full,3)) %>%
  select(country_name, ISO3c, cow_code, everything()) %>%
  select(-c(Work_Inj_Cov_LF, Sick_Cov_LF, Sick_Cov_Pop, RR_Sing, RR_Fam, RR_Avg_Sing, Dis_RR, gini_disp, gini_mkt, lfp_female, lfp_total, gdppc, pop, Region_factor, gdppc_k, workinjury_coverage_GEIP, workinjury_coverage_WSPR, Work_Inj_Cov_LF_imputed, RR_Sing_imputed, lfp_total_I, Work_Inj_Cov_LF_imputed_full, RR_Sing_imputed_full))
  


```


### Codebook

```{r codes}

gwip20_labels <- as.data.frame(colnames(gwip20))

colnames(gwip20_labels) <- "variables"

gwip20_labels <- gwip20_labels %>%
  mutate(label = case_when(
    variables == "country_name" ~ "Country name in English",
    variables == "cow_code" ~ "Correlates of War standard numeric country codes",
    variables == "ISO3c" ~ "International Standards Organization alpha-3 codes",
    variables == "independence" ~ "This variable gives information on the year a nation state becomes independent whether this arises out of association or end of colonial rule. If there were periods where the country lost independence but returned to its present or similar form (world wars or temporary annexation), then this code reflects the first instance except for cases of prolonged colonization, and having no work-injury laws prior to that colonization.",
    variables == "labor_workinjury_firstlaw" ~ "The year of introduction of the first law of any kind. It is the year the law is passed, not the moment it comes into force, so long as a timing of the law coming into force is written in the law or decided upon prior to its passing.",
    variables == "labor_workinjury_firstlaw_sspw" ~ "Taken from the SSPW 2017 country-specific html files using a text mining routine (Social Security Programs Throughout the World Data from the U.S. Social Security Administration). In our investigation of each country we often find different years than what is in the SSPW data. We include the SSPW data as of 2017 here for user reference as a variable. We do not argue that the SSPW is incorrect. They simply do not have a consistent coding scheme.",
    variables == "labor_workinjury_firstins" ~ "This is the year of the first law providing insurance against work-related injury. Insurance refers to the common definition, not the legal definition, as providing some form of shared protection against risk. That is why we refer to this as risk-pooling, not insurance, in the variable name. This variable indicates the first instance of mandatory social risk pooling. A provident fund, for example, counts as risk-pooling because workers have a shared pool of resources to protect them in case of an accident; again this is only by law and says nothing about the effectiveness or implementation of the law. Simply having a law requiring employer liability is not risk-pooling in our definition, because there is no insurance (fund, insurance policy, etc.) backing up the legal liability. If the first law of any form in a country was a risk-pooling law, then this variable is identical to labor_workinjury_firstlaw. If only a single occupation is covered but the state makes insurance mandatory for the entire single occupation (e.g., railroad workers or miners), then this would also count as the first risk-pooling law.",
    variables == "labor_workinjury_first_fund" ~ "First instance of risk pooling as a provident fund. It is missing if the country went directly from no law or employer liability without a fund to social insurance. Coding rules otherwise follow labor_workinjury_firstins. Provident fund means that paying money into a centralized fund is mandatory. This fund has the purpose of paying compensation.",
    variables == "labor_workinjury_first_socins" ~ "First instance of social insurance for any blue-collar worker segment as defined in labor_workinjury_firstins. It must be mandatory that employees get private or associational firm insurance policies, or the government itself is the insurer. The key difference in social insurance as a form of risk-pooling is that fault is no longer on the burden of the employee. The employee is automatically entitled to accident compensation unless otherwise disproven by the insurer or other interested party - this shifts fault away from the employee by default.",
    variables == "labor_workinjury_firstnat" ~ "The year of the first law when the country exists as an independent political entity. Many first laws were implemented under colonial or foreign rule. If the law was simply carried over from a previous law before independence then the year is identical to the year of independence. In cases where the first national law is passed within 2 years prior to independence, we code the year of the passing of the law rather than the year of independence and code labor_workinjury_firstnat_carriedover = no, as in Grenada, Bulgaria, Iceland, Benin, Cameroon and Samoa.",
    variables == "labor_workinjury_firstnat_carriedover" ~ "This variable indicates, whether a law from colonial, foreign or imperial rule is carried over after independence. The exceptions are states that became nation states and carried the law over and did not have a geopolitical change (i.e., boundaries stayed the same); for these cases the first national law could come while not a fully autonomous nation state",
    variables == "labor_workinjury_firstlaw_programtype" ~ "The program type of a work-injury law can either be **Risk-Pooling**, meaning that there is some kind of mandatory risk pooling or **Employer liability** meaning that there is no involvement of the state in financing and risk pooling, only an obligation for the employers to compensate in the case of work accidents. In this version of the variable, **Risk-Pooling** refers to either provident funds set up by the state or made mandatory by the state or actual social insurance with a private or state carrier (as with the coding of labor_workinjury_firstins). **Employer liability** is not the same as common or civil law codes that assign fault in case of injury. **Employer liability** means that there is a specific law targeting employers in industrial accidents and deaths as something more than basic common and civil codes. This is a variable constructed from a combination of labor_workinjury_firstlaw, labor_workinjury_first_fund, and labor_workinjury_first_socins",
    variables == "labor_workinjury_firstlaw_bluecollar_fullcoverage" ~ "This is the year of full coverage for all blue-collar groups: manual labor; excluding agriculture and domestic workers. All means at least two-thirds, or all the major sectors of industry; exceptions should not be occupation or industry-specific. We code this as covering all as long as all enterprises with at least 5 or more employees are covered.",
    variables == "workinjury_coverage_ILO" ~ "Coverage rates as the mean of the ILO's Global Employment Injury Protection study of 2014, and World Social Protection Report of 2021",
    variables == "replacement_rate_perm_sspw" ~ "The average replacement rate after a work-injury, for a permanent or apparently permanent disability for the first 12 months of incapacity, rate for 2020 or latest available, source = GWIP 2020",
    variables == "workinjury_duration_perm_2020" ~ "Duration of benefits calculated in replacement_rate_perm, under the assumption that the worker has or appears to have a permanent disability, source = GWIP 2020.",
    variables == "replacement_rate_temp_sspw" ~ "The average replacement rate after a work-injury, from previous income for the first 6 months of incapacity to work in 2020 or latest available, source = GWIP 2020",
    variables == "workinjury_duration_temp_2020" ~ "Duration of the temporary benefits calculated in replacement_rate_temp, source = GWIP 2020",
    variables == "Region" ~ "UN Sub-Region, derived from R package countrycode",
    variables == "global_south" ~ "A rough categorization of the world into Global North (colonizer) and South (colonized) countries, without treatment of China and Russia as colonizers",
    variables == "colonial_history" ~ "A rough categorization of the former colonies. When multiple colonizers attempt made to select the primary (longest, strongest). This includes Russia's pre-1922 colonies",
    variables == "workinjury_coverage_SIED_harmo" ~ "The SIED variable acovratl, coverage of workers under work-injury law as a percentage of the labor force with missing values harmonized from ILO GEIP coverage data, source = SIED 2020 & GEIP 2020",
    variables == "workinjury_coverage_SIED_harmo_full" ~ "The SIED variable acovratl, coverage of workers under work-injury law as a percentage of the labor force with missing values harmonized and imputed from ILO GEIP and GWIP variables, rate for 2020 or latest available, source = SIED 2020, GEIP 2020 & GWIP 2020",
    variables == "workinjury_replacement_rate_single_SIED_harmo" ~ "The SIED variable azrr26si, replacement rate after a work-injury for the first 6 months for a standard net single worker, harmonized from replacement_rate_temp, rate for 2020 or latest, source = SIED 2020 & GWIP 2020",
    variables == "workinjury_replacement_rate_single_SIED_harmo_full" ~ "The SIED variable azrr26si, replacement rate after a work-injury for the first 6 months for a standard net single worker, harmonized and imputed from replacement_rate_temp and socioeconomic/historical variables, rate for 2020 or latest, source = SIED 2020 & GWIP 2020"
  ))

sum_gwip <- gwip20 %>%
  summarise_all(mean, na.rm = T)

sd_gwip <- gwip20 %>%
  summarise_all(sd, na.rm = T)

min_gwip <- gwip20 %>%
  summarise_all(min, na.rm = T)

max_gwip <- gwip20 %>%
  summarise_all(max, na.rm = T)

nonna_gwip <- as.data.frame(t(sapply(gwip20 , function( x ) sum( !is.na( x ) ) )))

gwip_desc <- as.data.frame(t(rbind(sum_gwip, sd_gwip, min_gwip, max_gwip, nonna_gwip)))

colnames(gwip_desc) <- c("Mean", "SD", "Min", "Max", "N")

gwip_desc <- gwip_desc[!(rownames(gwip_desc) %in% c("country_name", "ISO3c", "cow_code", "labor_workinjury_firstlaw_programtype", "Region", "Colonial_Legacy", "workinjury_duration_perm_2020_sspw", "workinjury_duration_temp_2020_sspw","labor_workinjury_firstnat_carriedover")),]

gwip_desc[,1] <- as.numeric(gwip_desc[,1])
gwip_desc[,2] <- as.numeric(gwip_desc[,2])
gwip_desc[,3] <- as.numeric(gwip_desc[,3])
gwip_desc[,4] <- as.numeric(gwip_desc[,4])
gwip_desc[,5] <- as.numeric(gwip_desc[,5])

gwip_desc <- round(gwip_desc, 2)

write.csv(gwip_desc, here::here("Results", "gwip_desc.csv"))
```

### Save

```{r save}
write_csv(gwip20, here::here("data", "gwip20.csv"))

write_csv(gwip20_labels, here::here("data", "gwip20_labels.csv"))

```

## Colophon

```{r colophon}
sessionInfo()

```

## Citations

```{r citations}
packages %>%
  map(citation) %>%
  print(style = "text")
```
