==> 03_Analyses.Rmd <==
---
title: "03_Analyses"
author: "Nate Breznau"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = F, message = F)

packages <- c("tidyverse",
              "countrycode",
              "ggplot2",
              "ggrepel",
              "ggpubr",
              "sjPlot",
              "ragg",
              'rnaturalearth',
              'sf', 
              "ggmap",
              "psych")

pacman::p_load(packages, character.only = T)
  
  
gwip20 <- as.data.frame(read_csv(here::here("data", "gwip20.csv")))
gwip20_labels <- as.data.frame(read_csv(here::here("data", "gwip20_labels.csv")))
```

### Figure 1. GWIP Coverage Map

#### Setup

```{r fig1setup}
wi_map <- ne_countries(returnclass = "sf") %>%
  mutate(id = iso_a3,
         id = ifelse(id == "-99", iso_a3_eh, id)) %>%
  select(id, geometry)

# add Greenland as a row identical to Denmark for a better looking map
gwip20_GRL <- gwip20 %>%
  subset(ISO3c == "DNK") %>%
  mutate(ISO3c = "GRL")

gwip20_map <- rbind(gwip20, gwip20_GRL) %>%
  mutate(id = ISO3c)

wi_map <- full_join(wi_map, gwip20_map, by = c("id")) %>%
  subset(id != "ATA")

rm(gwip20_GRL)
```


#### Plot Fig

```{r fig1plot}
agg_png(here::here("results", "Fig1.png"), width = 1000, height = 800, res = 144)

map1 <- wi_map %>%
  mutate(y = 100*workinjury_coverage_SIED_harmo_full) %>%
  ggplot() +
  geom_sf(aes(fill = y), size = 0.15) +
    scale_fill_binned(type = "viridis",
                      direction = -1,
                      #n.breaks = 7,
                      na.value = "grey90") +
    coord_sf(label_axes = "----") +
    labs(fill = "Coverage\nas % of\nLabor Force", size = 2) +
    theme_classic() +
  theme(
    legend.position = "bottom"
  )


map1

dev.off()

knitr::include_graphics(here::here("results", "Fig1.png"))
```

### Figure 2. RR Map

```{r fig7}
agg_png(here::here("results", "Fig2.png"), width = 1000, height = 800, res = 144)

map2 <- wi_map %>%
  mutate(y = 100*workinjury_replacement_rate_single_SIED_harmo_full) %>%
  ggplot() +
  geom_sf(aes(fill = y), size = 0.15) +
    scale_fill_binned(type = "viridis",
                      breaks = c(60,80,100,110),
                      direction = -1,
                      #n.breaks = 7,
                      na.value = "grey90") +
    coord_sf(label_axes = "----") +
    labs(fill = "Replacement Rate\n26-Week Work Incapacity", size = 2) +
    theme_classic() +
  theme(
    legend.position = "bottom"
  )


map2

dev.off()

knitr::include_graphics(here::here("results", "Fig2.png"))
```

### Variation in 2020

#### Percent in Agriculture

World Bank 2024

https://data.worldbank.org/indicator/SL.AGR.EMPL.ZS?end=2017&most_recent_year_desc=false&start=2014

World Bank. 2024. ‘Employment in Agriculture (% of Total Employment) 2017 (Modeled ILO Estimate)’.


```{r wb}
agdata <- read_csv(here::here("data", "percent_ag_wb.csv")) %>%
  mutate(ISO3c = countrycode::countrycode(Country, "country.name", "iso3c"),
         ISO3c = ifelse(Country == "Turkiye", "TUR", ISO3c),
         ag_pct = `Most Recent Value`) %>%
  subset(!is.na(ISO3c), select = c(ISO3c, ag_pct))

gwip20 <- gwip20 %>%
  left_join(agdata, by = "ISO3c") 
```
##### Inverse Cov RR

There is an inverse association between coverage and replacement rates.

```{r inv_cov}
r_invcov <- round(cor(gwip20$workinjury_replacement_rate_single_SIED_harmo_full, gwip20$workinjury_coverage_SIED_harmo_full),3)

r_p <- corr.test(gwip20$workinjury_replacement_rate_single_SIED_harmo_full, gwip20$workinjury_coverage_SIED_harmo_full)$p

paste0("R = ", r_invcov, ", p =", round(r_p,3))
```

### Figure 3 & 4. Coverage by percent Agriculture

#### Global South

```{r fig8a}
agg_png(here::here("results", "Fig3.png"), res = 144, width = 1200, height = 1000)

gwip20 %>%
  subset(global_south == 1) %>%
  mutate(ag_pct_non = 100-ag_pct,
         ag_pct_non = ifelse(ag_pct_non < 25, 25, ag_pct_non)) %>%
  ggplot(aes(ag_pct_non, workinjury_coverage_SIED_harmo_full, label = ISO3c)) +
  geom_smooth(method = "lm", color = "#472D7BFF", size = 0.4) +
  geom_text_repel(max.overlaps = Inf) +
    stat_cor(aes(ag_pct_non, workinjury_coverage_SIED_harmo_full),
           color = "#472D7BFF",
           label.x = 88,
           label.y = 0.05) +
  xlim(25,100) +
  ylim(0,1) +
  xlab("Percentage of Labor Force\nNot in Agriculture") +
  ylab("Work-Injury Coverage as a Percentage\nof the Labor Force") +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results", "Fig3.png"))
```
#### Global North

```{r fig8b}
agg_png(here::here("results", "Fig4.png"), res = 144, width = 1200, height = 1000)

gwip20 %>%
  subset(global_south == 0) %>%
  mutate(ag_pct_non = 100-ag_pct,
         ag_pct_non = ifelse(ag_pct_non < 25, 25, ag_pct_non)) %>%
  ggplot(aes(ag_pct_non, workinjury_coverage_SIED_harmo_full, label = ISO3c)) +
  geom_smooth(method = "lm", color = "#472D7BFF", size = 0.4) +
  geom_text_repel(max.overlaps = Inf) +
  xlim(90,100) +
  ylim(0.65,1) +
  stat_cor(aes(ag_pct_non, workinjury_coverage_SIED_harmo_full),
           color = "#472D7BFF",
           label.x = 97,
           label.y = 0.67) +
    xlab("Percentage of Labor Force\nNot in Agriculture") +
  ylab("Work-Injury Coverage as a Percentage\nof the Labor Force") +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results", "Fig4.png"))
```
### Institutions


#### Individualism data

https://osf.io/knvzy

Santos, Henri C., Michael E. W. Varnum, and Igor Grossmann. 2017. ‘Global Increases in Individualism’. Psychological Science 28(9):1228–39. doi: 10.1177/0956797617700622.


```{r indv}
CCHANGE <- read_csv(here::here("data", "Global Increases in Individualism Data.csv"))

# we run their R code for compiling the mean scores across countries
source(here::here("data", "Santos_script.R"))

indiv <- CCHANGE %>%
  mutate(ISO3c = countrycode(Country, "country.name", "iso3c"),
         individualism = Values.raw) %>%
  select(ISO3c, individualism) %>%
  group_by(ISO3c) %>%
  summarise_all(mean, na.rm = T) %>%
  subset(!is.na(individualism))

gwip20 <- gwip20 %>%
  left_join(indiv, by = "ISO3c")
  
```

#### Fig 5. Coverage and Year

```{r fig10}
agg_png(here::here("results", "Fig5.png"), res = 144, width = 1200, height = 800)

gwip20 %>%
  mutate(duration = 2020 - labor_workinjury_firstlaw) %>%
  ggplot(aes(duration, workinjury_coverage_SIED_harmo_full)) +
  geom_smooth(method = "lm", color = "#472D7BFF", size = 0.4) +
  geom_point() +
  #xlim(90,100) +
  #ylim(0.65,1) +
  stat_cor(aes(duration, workinjury_coverage_SIED_harmo_full),
           color = "#472D7BFF",
           label.x = 130,
           label.y = 0.1
           ) +
  xlab("Years Since the Introduction\nof First Work-Injury Law") +
  ylab("Work-Injury Coverage as a Percentage\nof the Labor Force") +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results", "Fig5.png"))
```


#### Fig 6. Replacement Rates and Individualism

```{r fig11}
agg_png(here::here("results", "Fig6.png"), res = 144, width = 1200, height = 900)

gwip20 %>%
  ggplot(aes(individualism, workinjury_replacement_rate_single_SIED_harmo_full, label = ISO3c)) +
  geom_smooth(method = "lm", color = "#472D7BFF", size = 0.4) +
  geom_text_repel(max.overlaps = Inf) +
  #xlim(90,100) +
  ylim(0.51,1.37) +
  stat_cor(aes(individualism, workinjury_replacement_rate_single_SIED_harmo_full),
           color = "#472D7BFF",
           label.x = -1.4,
           label.y = 0.52) +
  xlab("Cultural Individualism Values, Averaged") +
  ylab("Work-Injury Replacement Rate\n26-Weeks Incapacity") +
  theme_classic()

dev.off()

knitr::include_graphics(here::here("results", "Fig6.png"))
```

## Colophon

```{r colophon}
sessionInfo()

```

## Citations

```{r citations}
packages %>%
  map(citation) %>%
  print(style = "text")
```
